refactor(pool-page): extract _tabs/rebalances-tab
Move RebalancesTab (~310 lines incl. p90/p95 reward outlier
highlighting from PR #259) out of page.tsx into _tabs/rebalances-tab.tsx
along with its helpers:
- REWARD_OUTLIER_TIERS constant
- REWARD_HIST_REFRESH_MS constant
- RewardThresholds type
- computeRewardThresholds (exported)
- renderRewardCell (exported)

page.tsx re-exports computeRewardThresholds + renderRewardCell so
__tests__/reward-outliers.test.tsx (which imports them from "../page")
keeps working without edits.

Behavior-preserving — verified via:
- pnpm test (1366/1366 green; reward-outliers.test.tsx and
  page.test.tsx Rebalances suites pass)
- pnpm typecheck (clean)
- chrome-devtools snapshot of /pool/<id>?tab=rebalances vs baseline:
  zero structural changes, console clean

page.tsx: 1257 → 894 lines (-363).
